+++ ChangeLog.md
@@ -18,6 +18,8 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ### Fixed
 
+- Fixed return value for sendCommand function.
+
 ***
 
 ## [1.3.0]